-Added dynamic map script (dynamic1.mapscript) -Refactored g/h/f values to int from float.

+++ src/agents/SlidingWindow.java
@@ -80,7 +80,7 @@ public class SlidingWindow
 		// TODO: probably a better interface than this, but as long as it is remembered to
 		// check first!
 
-		return (m_count > 0) ? m_sum / m_count : 0;
+		return (m_count > 0) ? (float)m_sum / m_count : 0;
 	}
 
 	// Used to trigger a "settling" state
